# product-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the product catalog and inventory management service for the e-commerce platform.",
        "title": "Product Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Septareno Nugroho Aji",
            "email": "renoaji25sep@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/categories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product category (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.Category"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.CategorySuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied: Admins only",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "category already exists",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "could not create category",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Get products with pagination and filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated products",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by product name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum price",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum price",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.PaginatedProducts"
                        }
                    },
                    "500": {
                        "description": "could not retrieve products",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body / missing required fields / invalid category ID",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied: Admins only",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "product already exists",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "could not create product",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Get a single product by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ProductDataResponse"
                        }
                    },
                    "400": {
                        "description": "invalid product ID",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "could not retrieve product",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update product details (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product update data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ProductSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied: Admins only",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "could not update product",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a product by ID (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid product ID",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied: Admins only",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "could not delete product",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "product-service_internal_domain.Category": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "products": {
                    "description": "Inverse association",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product-service_internal_domain.Product"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.CategorySuccessResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/product-service_internal_domain.Category"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.CreateProductRequest": {
            "type": "object",
            "required": [
                "category_ids",
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "category_ids": {
                    "description": "User only sends [1, 2, 3]",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "product-service_internal_domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.PaginatedProducts": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product-service_internal_domain.ProductResponse"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "product-service_internal_domain.Product": {
            "type": "object",
            "required": [
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "categories": {
                    "description": "Many-to-Many association",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product-service_internal_domain.Category"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.ProductDataResponse": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/product-service_internal_domain.ProductResponse"
                }
            }
        },
        "product-service_internal_domain.ProductResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product-service_internal_domain.CategoryResponse"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.ProductSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/product-service_internal_domain.ProductResponse"
                }
            }
        },
        "product-service_internal_domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "product-service_internal_domain.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "category_ids": {
                    "description": "If provided, we replace all categories",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}